requirements.txt
bootstrap4~=0.1.0
Flask~=3.1.1
requests~=2.32.3
python-dotenv~=1.1.0
